Finished the bulk of Game -> Run conversion

+++ src/structures/data/DataInstance.java
@@ -35,7 +35,7 @@ public class DataInstance {
         myVelocity = Vector.ZERO;
     }
 
-    public double getID() {
+    public long getID() {
         return myID;
     }
